Move cache tests into docker
